Since the dependency on the Cern colt library has been taken out, ProbMath now defines the Gamma and the Beta function for positive reals, as well as the gammaln function ln(gamma(x)).

+++ dsol-core/src/main/java/nl/tudelft/simulation/dsol/model/inputparameters/InputParameterDistDiscreteSelection.java
@@ -159,7 +159,7 @@ public class InputParameterDistDiscreteSelection extends InputParameterSelection
             Throw.when((Double) get("p").getValue() >= 1.0, InputParameterException.class, "DistBinomial: p >= 1.0");
             try
             {
-                this.dist = new DistBinomial(getStream(), (Long) get("n").getValue(), (Double) get("p").getValue());
+                this.dist = new DistBinomial(getStream(), (Integer) get("n").getValue(), (Double) get("p").getValue());
             }
             catch (Exception exception)
             {
@@ -298,7 +298,7 @@ public class InputParameterDistDiscreteSelection extends InputParameterSelection
             Throw.when((Double) get("p").getValue() >= 1.0, InputParameterException.class, "DistNegBinomial: p >= 1.0");
             try
             {
-                this.dist = new DistNegBinomial(getStream(), (Long) get("n").getValue(), (Double) get("p").getValue());
+                this.dist = new DistNegBinomial(getStream(), (Integer) get("n").getValue(), (Double) get("p").getValue());
             }
             catch (Exception exception)
             {
